remove legacy apigw provider
